slices-to-3d-brain-vae/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py36_0
astor=0.7.1=py36_0
attrs=19.1.0=py36_1
backcall=0.1.0=pypi_0
blas=1.0=mkl
bleach=3.1.0=py36_0
c-ares=1.15.0=h7b6447c_1
ca-certificates=2020.1.1=0
cachetools=3.1.1=pypi_0
certifi=2019.11.28=py36_0
cffi=1.12.3=py36h2e261b9_0
cloudpickle=1.2.1=py_0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cupti=10.0.130=0
cycler=0.10.0=py36_0
cytoolz=0.10.0=py36h7b6447c_0
dask-core=2.3.0=py_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py36_1
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
future=0.17.1=pypi_0
gast=0.2.2=py36_0
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
google-pasta=0.1.7=py_0
grpcio=1.16.1=py36hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.9.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
imageio=2.5.0=py36_0
intel-openmp=2019.4=243
ipdb=0.12=pypi_0
ipykernel=5.1.2=py36h39e3cac_0
ipython=7.6.1=pypi_0
ipython-genutils=0.2.0=pypi_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.5.1=py_0
jedi=0.14.1=pypi_0
jinja2=2.10.1=py36_0
joblib=0.13.2=py36_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.2=py36_0
jupyter=1.0.0=py36_7
jupyter_client=5.3.1=py_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.5.0=py_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.8.0=hd408876_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py36_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.1.0=py36h5429711_0
medpy=0.4.0=pypi_0
mistune=0.8.4=py36h7b6447c_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.12=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
natsort=7.0.1=py_0
nbconvert=5.5.0=py_0
nbformat=4.4.0=py36_0
ncurses=6.1=he6710b0_1
networkx=2.3=py_0
nibabel=2.5.0=pypi_0
nilearn=0.6.2=pypi_0
ninja=1.9.0=py36hfd86e86_0
notebook=6.0.0=py36_0
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
olefile=0.46=py36_0
openssl=1.1.1d=h7b6447c_4
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.5.1=py_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=pypi_0
pillow=6.1.0=py36h34e0f95_0
pip=19.1.1=py36_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.9=py36_0
protobuf=3.8.0=py36he6710b0_0
ptyprocess=0.6.0=pypi_0
pycparser=2.19=py36_0
pygments=2.4.2=py_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.14.11=py36h7b6447c_0
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py36_0
pytz=2019.1=py_0
pywavelets=1.0.3=py36hdd07704_1
pyzmq=18.1.0=py36he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.5.4=py_0
readline=7.0=h7b6447c_5
scikit-image=0.15.0=py36he6710b0_0
scikit-learn=0.21.3=py36hd81dba3_0
scipy=1.2.1=py36h7c811a0_0
send2trash=1.5.0=py36_0
setuptools=41.0.1=py36_0
simpleitk=1.2.2=pypi_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sklearn=0.0=pypi_0
sqlite=3.29.0=h7b6447c_0
tb-nightly=1.15.0a20190806=pypi_0
tensorboard=1.14.0=py36hf484d3e_0
tensorflow=1.14.0=gpu_py36h57aa796_0
tensorflow-base=1.14.0=gpu_py36h8d69cac_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py36_1
terminado=0.8.2=py36_0
testpath=0.4.2=py36_0
tk=8.6.8=hbc83047_0
toolz=0.10.0=py_0
torch=1.2.0=pypi_0
torchvision=0.4.0=pypi_0
tornado=6.0.3=py36h7b6447c_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=pypi_0
webencodings=0.5.1=py36_1
werkzeug=0.15.4=py_0
wheel=0.33.4=py36_0
widgetsnbextension=3.5.1=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.4=h14c3975_4
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
